# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lilypad API",
        "description": "API for lilypad",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/inventory": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get inventory items by Part Number or Location Group",
                "description": "This endpoint retrieves inventory items associated with a specific part number, location group, or all inventory items if no filters are specified.",
                "operationId": "f36f846f5144a5814a75101323bdd665",
                "parameters": [
                    {
                        "name": "num",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "PN-12345"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Group A"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "num": {
                                        "description": "The number of the part.",
                                        "type": "string",
                                        "example": "PN-12345"
                                    },
                                    "name": {
                                        "description": "The name of the location group.",
                                        "type": "string",
                                        "example": "Group A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventory items retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message.",
                                            "type": "string",
                                            "example": "Inventory items retrieved successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message.",
                                            "type": "string",
                                            "example": "Invalid parameters. Only one filter is allowed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message.",
                                            "type": "string",
                                            "example": "No inventory found for the specified filters."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Store a new inventory item",
                "description": "This endpoint allows the user to create a new inventory item or update an existing one based on the part number and location.",
                "operationId": "5ca8e6a52f4241a17ccfef607c7252be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "PartNumber": {
                                        "description": "The unique number of the part.",
                                        "type": "string",
                                        "example": "10010"
                                    },
                                    "PartDescription": {
                                        "description": "A brief description of the part.",
                                        "type": "string",
                                        "example": "High-quality widget"
                                    },
                                    "Location": {
                                        "description": "The location where the part is stored.",
                                        "type": "string",
                                        "example": "Main"
                                    },
                                    "Qty": {
                                        "description": "The quantity of the part being added to inventory.",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "UOM": {
                                        "description": "The unit of measure for the part.",
                                        "type": "string",
                                        "example": "Kilogram"
                                    },
                                    "Cost": {
                                        "description": "The cost of the part.",
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "QbClass": {
                                        "description": "The QuickBooks class for the part.",
                                        "type": "string",
                                        "example": "Sales"
                                    },
                                    "Date": {
                                        "description": "The date when the inventory entry is created.",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-17"
                                    },
                                    "Note": {
                                        "description": "Any additional notes regarding the inventory.",
                                        "type": "string",
                                        "example": "New shipment received"
                                    },
                                    "TrackingType": {
                                        "description": "The type of tracking used for the part.",
                                        "type": "string",
                                        "example": "Lot Number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inventory Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message indicating the inventory was created.",
                                            "type": "string",
                                            "example": "Inventory Created Successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Inventory has been updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message indicating the inventory was updated.",
                                            "type": "string",
                                            "example": "Inventory has been updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating a validation issue.",
                                            "type": "string",
                                            "example": "UOM not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating the requested resource was not found.",
                                            "type": "string",
                                            "example": "Part Tracking not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}